Move the Id check screen from the cheatcodes folder to the signup folder
